ftrace: fix another cast
Fix a cast in the function profiler code that will be enabled shortly.

Cast to uintptr_t to preserve the capability metadata.

+++ kernel/trace/ftrace.c
@@ -458,7 +458,7 @@ function_stat_next(void *v, int idx)
 	struct ftrace_profile *rec = v;
 	struct ftrace_profile_page *pg;
 
-	pg = (struct ftrace_profile_page *)((unsigned long)rec & PAGE_MASK);
+	pg = (struct ftrace_profile_page *)((uintptr_t)rec & PAGE_MASK);
 
  again:
 	if (idx != 0)
